rename the reader file
